requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
openai==2.8.1
pydantic==2.5.0
python-dateutil==2.8.2
email-validator==2.1.0
phonenumbers==8.13.25
python-dotenv==1.0.0

